added commented working version
